Validate static preference on save
Verify the presence and validity of static preferences added by the
user, as previously implemented for custom preferences via the admin
panel.

From now on, whether custom preferences or form preferences are selected
in the admin panel, a validation check will be performed upon saving
to ensure their accuracy. This will assist the user with the setup
process.